chore(ndm): regenerate the discovery bazel target and tidy the probe cut
Gazelle had not run since the probe files replaced credentials.go and
pingcapability.go, so the bazel target listed four deleted files and none
of the new ones. Also move statusString next to the probe interfaces that
use it, compute the range size in int64 so a wide prefix cannot overflow
int on a 32-bit build, and drop a duplicate package comment.

+++ comp/ndmdiscovery/impl/chunks.go
@@ -3,7 +3,6 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2026-present Datadog, Inc.
 
-// Package ndmdiscoveryimpl implements the ndmdiscovery component.
 package ndmdiscoveryimpl
 
 import (
@@ -40,14 +39,14 @@ func newChunkPlan(cidr string, ignored []string, maxAddresses int) (*chunkPlan, 
 	}
 	prefix = prefix.Masked()
 
-	total := 1 << (32 - prefix.Bits())
-	if total > maxAddresses {
+	total := int64(1) << (32 - prefix.Bits())
+	if total > int64(maxAddresses) {
 		return nil, fmt.Errorf("range %q holds %d addresses, which exceeds the maximum of %d", cidr, total, maxAddresses)
 	}
 
 	p := &chunkPlan{
 		prefix:  prefix,
-		total:   total,
+		total:   int(total),
 		ignored: make(map[netip.Addr]struct{}, len(ignored)),
 	}
 	for _, raw := range ignored {
